Restructure unit tests
