+++ Innofactor.EfCoreJsonValueConverter.Test/Components/TestableDbContext.cs
@@ -1,0 +1,22 @@
+﻿using Innofactor.EfCoreJsonValueConverter.Test.Entities;
+using Microsoft.EntityFrameworkCore;
+
+namespace Innofactor.EfCoreJsonValueConverter.Test.Components
+{
+  public class TestDbContext : DbContext
+  {
+    public DbSet<Customer> Customers { get; set; }
+
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder) {
+      base.OnConfiguring(optionsBuilder);
+      optionsBuilder.UseInMemoryDatabase("Test");
+    }
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder) {
+      modelBuilder.Entity<Customer>(c => {
+        c.Property(e => e.Address).HasJsonValueConversion();
+        c.Property(e => e.Address2).HasJsonValueConversion();
+      });
+    }
+  }
+}
