+++ Innofactor.EfCoreJsonValueConverter.Test/Entities/Customer.cs
@@ -1,0 +1,19 @@
+﻿namespace Innofactor.EfCoreJsonValueConverter.Test.Entities
+{
+  public class Customer
+  {
+    public int Id { get; set; }
+
+    [JsonField]
+    public Address Address { get; set; }
+
+    [JsonField]
+    public AddressWithEquality Address2 { get; set; }
+
+  }
+
+  public class CustomerWithPlainField
+  {
+    private string name;
+  }
+}
