+++ Innofactor.EfCoreJsonValueConverter.Test/Innofactor.EfCoreJsonValueConverter.Test.csproj
@@ -124,6 +124,9 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Components\TestableDbContext.cs" />
+    <Compile Include="Entities\Customer.cs" />
+    <Compile Include="Entities\Address.cs" />
     <Compile Include="ModelBuilderExtensionsTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
